* [core]: fix some more clang warnings
* tautologocal pointer compare
* implicite enum conversion
...

+++ src/templates/objects_profile/oyProfileTag_s_.template.c
@@ -35,6 +35,6 @@
   if((int)type >= oyNAME_DESCRIPTION)
     sprintf( &text[strlen(text)], "%s %s\noffset: %lu size: %lu/%lu",
              oyICCTagDescription(s->use), oyICCTagTypeName(s->tag_type_),
-             s->offset_orig, s->size_, s->size_check_
+             (long unsigned int)s->offset_orig, (long unsigned int)s->size_, (long unsigned int)s->size_check_
            );
 {% endblock %}
