+++ src/templates/objects_value/oyBlob_s_.template.c
@@ -10,11 +10,11 @@
   } else
   if(type == oyNAME_NAME)
     sprintf( &text[strlen(text)], "%s %lu",
-             s->type, s->size
+             s->type, (long unsigned int)s->size
            );
   else
   if((int)type >= oyNAME_DESCRIPTION)
     sprintf( &text[strlen(text)], "type: %s size: %lu flags: %d",
-             s->type, s->size, s->flags
+             s->type, (long unsigned int)s->size, s->flags
            );
 {% endblock %}
